Germany's Economic Development from 2015 to 2019
Machine Learning Software Predicting Germany's Economic Development from 2015 to 2019

I developed Machine Learning Software that Predicts Germany's Economic Development from 2015 to 2019. In this software, the purchasing power of the Euro against the US Dollar, its comparison against the US Dollar, and various economic parameters are estimated and classified. In addition, if the value of these data in 2019 is greater or less than the value in 2015, it is divided into various classes. Additionally, the data is visualized using the matplotlib library. Finally, the software and the model have been prepared with great care and the data has been entered manually and is completely original.